Adds workflow for tagging Terraform modules
Implements a new GitHub Actions workflow that automates the process of tagging Terraform modules based on the version specified in the config file.

This workflow checks out the repository, installs yq, configures the Git user, creates an annotated tag with a "v" prefix, and pushes the tag to the remote repository.